chore(statics): move tokenConfigs to statics
-moves types & constants from bitgo to statics

Ticket: BG-49378
