+++ modules/statics/src/index.ts
@@ -2,6 +2,7 @@ export * from './base';
 export * from './coins';
 export * from './networks';
 export * from './errors';
+export * from './tokenConfig';
 export { OfcCoin } from './ofc';
 export { UtxoCoin } from './utxo';
 export {
