+++ modules/statics/src/tokenConfig.ts
@@ -1,0 +1,323 @@
+import _ = require('lodash');
+import { Erc20Coin, StellarCoin, CeloCoin, EosCoin, AvaxERC20Token, AlgoCoin, SolCoin } from './account';
+import { CoinKind } from './base';
+import { coins } from './coins';
+import { Networks, NetworkType } from './networks';
+import { OfcCoin } from './ofc';
+
+export interface BaseTokenConfig {
+  coin: string;
+  decimalPlaces: number;
+  name: string;
+  type: string;
+}
+
+export interface BaseNetworkConfig extends BaseTokenConfig {
+  network: string;
+}
+
+export interface BaseContractAddressConfig extends BaseNetworkConfig {
+  tokenContractAddress: string;
+}
+
+export type AvaxcTokenConfig = BaseContractAddressConfig;
+export type CeloTokenConfig = BaseContractAddressConfig;
+export type EosTokenConfig = BaseContractAddressConfig;
+export type Erc20TokenConfig = BaseContractAddressConfig;
+export type StellarTokenConfig = BaseNetworkConfig;
+
+export type SolTokenConfig = BaseNetworkConfig & {
+  tokenAddress: string;
+};
+
+export type AlgoTokenConfig = BaseNetworkConfig & {
+  alias?: string;
+};
+
+export type OfcTokenConfig = BaseTokenConfig & {
+  backingCoin: string;
+  isFiat: boolean;
+};
+
+export interface Tokens {
+  bitcoin: {
+    eth: {
+      tokens: Erc20TokenConfig[];
+    };
+    xlm: {
+      tokens: StellarTokenConfig[];
+    };
+    algo: {
+      tokens: AlgoTokenConfig[];
+    };
+    ofc: {
+      tokens: OfcTokenConfig[];
+    };
+    celo: {
+      tokens: CeloTokenConfig[];
+    };
+    eos: {
+      tokens: EosTokenConfig[];
+    };
+    avaxc: {
+      tokens: AvaxcTokenConfig[];
+    };
+    sol: {
+      tokens: SolTokenConfig[];
+    };
+  };
+  testnet: {
+    eth: {
+      tokens: Erc20TokenConfig[];
+    };
+    xlm: {
+      tokens: StellarTokenConfig[];
+    };
+    algo: {
+      tokens: AlgoTokenConfig[];
+    };
+    ofc: {
+      tokens: OfcTokenConfig[];
+    };
+    celo: {
+      tokens: CeloTokenConfig[];
+    };
+    eos: {
+      tokens: EosTokenConfig[];
+    };
+    avaxc: {
+      tokens: AvaxcTokenConfig[];
+    };
+    sol: {
+      tokens: SolTokenConfig[];
+    };
+  };
+}
+
+// Get the list of ERC-20 tokens from statics and format it properly
+const formattedErc20Tokens = coins.reduce((acc: Erc20TokenConfig[], coin) => {
+  if (coin instanceof Erc20Coin) {
+    let baseCoin: string;
+    switch (coin.network) {
+      case Networks.main.ethereum:
+        baseCoin = 'eth';
+        break;
+      case Networks.test.kovan:
+        baseCoin = 'teth';
+        break;
+      case Networks.test.goerli:
+        baseCoin = 'gteth';
+        break;
+      default:
+        throw new Error(`Erc20 token ${coin.name} has an unsupported network`);
+    }
+
+    acc.push({
+      type: coin.name,
+      coin: baseCoin,
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      tokenContractAddress: coin.contractAddress.toString().toLowerCase(),
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+export const ethGasConfigs = {
+  minimumGasPrice: 1000000000, // minimum gas price a user can provide (1 Gwei)
+  defaultGasPrice: 20000000000, // default gas price if estimation fails (20 Gwei)
+  maximumGasPrice: 2500000000000, // minimum gas price a user can provide (2500 Gwei)
+  defaultGasLimit: 500000, // Default gas limit we set for contract send
+  defaultGasLimitTokenSend: 1000000, // Default gas limit we set for token send
+  minimumGasLimit: 30000, // minimum gas limit a user can set for a send
+  maximumGasLimit: 20000000, // Customers cannot set gas limits beyond this amount
+};
+// Get the list of Stellar tokens from statics and format it properly
+const formattedStellarTokens = coins.reduce((acc: StellarTokenConfig[], coin) => {
+  if (coin instanceof StellarCoin) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'xlm' : 'txlm',
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+// Get the list of OFC tokens from statics and format it properly
+const formattedOfcCoins = coins.reduce((acc: OfcTokenConfig[], coin) => {
+  if (coin instanceof OfcCoin) {
+    acc.push({
+      type: coin.name,
+      coin: 'ofc',
+      backingCoin: coin.asset,
+      name: coin.fullName,
+      decimalPlaces: coin.decimalPlaces,
+      isFiat: coin.kind === CoinKind.FIAT,
+    });
+  }
+  return acc;
+}, []);
+
+const formattedCeloTokens = coins.reduce((acc: CeloTokenConfig[], coin) => {
+  if (coin instanceof CeloCoin) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'celo' : 'tcelo',
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      tokenContractAddress: coin.contractAddress.toString().toLowerCase(),
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+const formattedEosTokens = coins.reduce((acc: EosTokenConfig[], coin) => {
+  if (coin instanceof EosCoin) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'eos' : 'teos',
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      tokenContractAddress: coin.contractName.toString().toLowerCase(),
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+const formattedAvaxCTokens = coins.reduce((acc: AvaxcTokenConfig[], coin) => {
+  if (coin instanceof AvaxERC20Token) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'avaxc' : 'tavaxc',
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      tokenContractAddress: coin.contractAddress.toString().toLowerCase(),
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+const formattedSolTokens = coins.reduce((acc: SolTokenConfig[], coin) => {
+  if (coin instanceof SolCoin) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'sol' : 'tsol',
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      tokenAddress: coin.tokenAddress,
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+export const formattedAlgoTokens = coins.reduce((acc: AlgoTokenConfig[], coin) => {
+  if (coin instanceof AlgoCoin) {
+    acc.push({
+      type: coin.name,
+      coin: coin.network.type === NetworkType.MAINNET ? 'algo' : 'talgo',
+      alias: coin.alias,
+      network: coin.network.type === NetworkType.MAINNET ? 'Mainnet' : 'Testnet',
+      name: coin.fullName,
+      decimalPlaces: coin.decimalPlaces,
+    });
+  }
+  return acc;
+}, []);
+
+export const tokens: Tokens = {
+  // network name for production environments
+  bitcoin: {
+    eth: {
+      tokens: formattedErc20Tokens.filter((token) => token.network === 'Mainnet'),
+    },
+    xlm: {
+      tokens: formattedStellarTokens.filter((token) => token.network === 'Mainnet'),
+    },
+    algo: {
+      tokens: formattedAlgoTokens.filter((token) => token.network === 'Mainnet'),
+    },
+    ofc: {
+      tokens: formattedOfcCoins.filter((token) => coins.get(token.type).network.type === NetworkType.MAINNET),
+    },
+    celo: {
+      tokens: formattedCeloTokens.filter((token) => token.network === 'Mainnet'),
+    },
+    eos: {
+      tokens: formattedEosTokens.filter((token) => token.network === 'Mainnet'),
+    },
+    avaxc: {
+      tokens: formattedAvaxCTokens.filter((token) => token.network === 'Mainnet'),
+    },
+    sol: {
+      tokens: formattedSolTokens.filter((token) => token.network === 'Mainnet'),
+    },
+  },
+  // network name for test environments
+  testnet: {
+    eth: {
+      tokens: formattedErc20Tokens.filter((token) => token.network === 'Testnet'),
+    },
+    xlm: {
+      tokens: formattedStellarTokens.filter((token) => token.network === 'Testnet'),
+    },
+    algo: {
+      tokens: formattedAlgoTokens.filter((token) => token.network === 'Testnet'),
+    },
+    ofc: {
+      tokens: formattedOfcCoins.filter((token) => coins.get(token.type).network.type === NetworkType.TESTNET),
+    },
+    celo: {
+      tokens: formattedCeloTokens.filter((token) => token.network === 'Testnet'),
+    },
+    eos: {
+      tokens: formattedEosTokens.filter((token) => token.network === 'Testnet'),
+    },
+    avaxc: {
+      tokens: formattedAvaxCTokens.filter((token) => token.network === 'Testnet'),
+    },
+    sol: {
+      tokens: formattedSolTokens.filter((token) => token.network === 'Testnet'),
+    },
+  },
+};
+
+/**
+ * Verify mainnet or testnet tokens
+ * @param tokens
+ */
+const verifyTokens = function (tokens: BaseTokenConfig[]) {
+  const verifiedTokens: Record<string, boolean> = {};
+  _.forEach(tokens, function (token) {
+    if (verifiedTokens[token.type]) {
+      throw new Error('token : ' + token.type + ' duplicated.');
+    }
+    verifiedTokens[token.type] = true;
+
+    if (
+      (token as BaseContractAddressConfig).tokenContractAddress &&
+      (token as BaseContractAddressConfig).tokenContractAddress !==
+        _.toLower((token as BaseContractAddressConfig).tokenContractAddress)
+    ) {
+      throw new Error(
+        'token contract: ' + token.type + ' is not all lower case: ' + (token as BaseContractAddressConfig)
+      );
+    }
+  });
+  return verifiedTokens;
+};
+
+const mainnetErc20Tokens = verifyTokens(tokens.bitcoin.eth.tokens);
+const mainnetStellarTokens = verifyTokens(tokens.bitcoin.xlm.tokens);
+export const mainnetTokens = _.assign({}, mainnetErc20Tokens, mainnetStellarTokens);
+
+const testnetErc20Tokens = verifyTokens(tokens.testnet.eth.tokens);
+const testnetStellarTokens = verifyTokens(tokens.testnet.xlm.tokens);
+export const testnetTokens = _.assign({}, testnetErc20Tokens, testnetStellarTokens);
